+++ source/extensions/omni.isaac.sensor/nodes/OgnIsaacComputeRTXRadarPointCloud.cpp
@@ -85,9 +85,33 @@ public:
         {
             return true;
         }
-
-
-        getTransformFromSensorPose(helper.m_gmo.frameEnd, matrixOutput);
+        omni::sensors::FrameAtTime frameEnd;
+#define RADAR_HAS_NO_TRANSFORM
+        // When the GMO output for the radar starts outputing the correct transform, you can
+        // undefine this.
+#ifndef RADAR_HAS_NO_TRANSFORM
+        frameEnd = helper.m_gmo.frameEnd;
+#else
+        pxr::GfVec3d translate{ 0, 0, 0 };
+        pxr::GfQuatd rotate{ 1, 0, 0, 0 };
+        pxr::UsdAttribute transAttr = omni::isaac::utils::getCameraAttributeFromRenderProduct(
+            "xformOp:translate", db.tokenToString(db.inputs.renderProductPath()));
+        pxr::UsdAttribute rotAttr = omni::isaac::utils::getCameraAttributeFromRenderProduct(
+            "xformOp:orient", db.tokenToString(db.inputs.renderProductPath()));
+        if (transAttr.IsValid() && rotAttr.IsValid())
+        {
+            omni::isaac::utils::safeGetAttribute(transAttr, translate);
+            omni::isaac::utils::safeGetAttribute(rotAttr, rotate);
+        }
+        frameEnd.orientation.w = (float)rotate.GetReal();
+        frameEnd.orientation.x = (float)rotate.GetImaginary()[0];
+        frameEnd.orientation.y = (float)rotate.GetImaginary()[1];
+        frameEnd.orientation.z = (float)rotate.GetImaginary()[2];
+        frameEnd.posM.x = (float)translate[0];
+        frameEnd.posM.y = (float)translate[1];
+        frameEnd.posM.z = (float)translate[2];
+#endif
+        getTransformFromSensorPose(frameEnd, matrixOutput);
         RadarAuxiliaryData* aux = reinterpret_cast<RadarAuxiliaryData*>(helper.m_gmo.auxiliaryData);
 
         // for the radar data:
@@ -132,6 +156,9 @@ public:
         _DEF_OUT_VAR(materialId);
         _DEF_OUT_VAR(objectId);
 #undef _DEF_OUT_VAR
+        bool hasSemId = ((aux->filledAuxMembers & RadarAuxHas::SEM_ID) == RadarAuxHas::SEM_ID);
+        bool hasMatId = ((aux->filledAuxMembers & RadarAuxHas::MAT_ID) == RadarAuxHas::MAT_ID);
+        bool hasObjId = ((aux->filledAuxMembers & RadarAuxHas::OBJ_ID) == RadarAuxHas::OBJ_ID);
 
         for (uint32_t i = 0; i < outSize; ++i)
         {
@@ -142,16 +169,16 @@ public:
             // z = r_m
             // scalar = rcs_dbsm
             // aux->rv_ms = rv_ms
-#define _ASSIGN_OUT(outputName, src) db_outputs_##outputName[i] = src[i]
-            _ASSIGN_OUT(radialDistance, helper.m_gmo.elements.z);
-            _ASSIGN_OUT(radialVelocity, aux->rv_ms);
-            _ASSIGN_OUT(azimuth, helper.m_gmo.elements.x);
-            _ASSIGN_OUT(elevation, helper.m_gmo.elements.y);
-            _ASSIGN_OUT(rcs, helper.m_gmo.elements.scalar);
-            _ASSIGN_OUT(semanticId, aux->semId);
-            _ASSIGN_OUT(materialId, aux->matId);
-            _ASSIGN_OUT(objectId, aux->objId);
-#undef _ASSIGN_OUT
+#define _IF_ASSIGN_OUT(has, outputName, src) db_outputs_##outputName[i] = (has) ? src[i] : 0
+            _IF_ASSIGN_OUT(true, radialDistance, helper.m_gmo.elements.z);
+            _IF_ASSIGN_OUT(true, radialVelocity, aux->rv_ms);
+            _IF_ASSIGN_OUT(true, azimuth, helper.m_gmo.elements.x);
+            _IF_ASSIGN_OUT(true, elevation, helper.m_gmo.elements.y);
+            _IF_ASSIGN_OUT(true, rcs, helper.m_gmo.elements.scalar);
+            _IF_ASSIGN_OUT(hasSemId, semanticId, aux->semId);
+            _IF_ASSIGN_OUT(hasMatId, materialId, aux->matId);
+            _IF_ASSIGN_OUT(hasObjId, objectId, aux->objId);
+#undef _IF_ASSIGN_OUT
         }
 
         return true;